closed #4559: Wrong behavior of multi-touch on iphone5s/ipadAir (arm64 arch)

+++ cocos/2d/platform/CCGLViewProtocol.cpp
@@ -37,7 +37,7 @@ namespace {
     static Touch* g_touches[EventTouch::MAX_TOUCHES] = { nullptr };
     static unsigned int g_indexBitsUsed = 0;
     // System touch pointer ID (It may not be ascending order number) <-> Ascending order number from 0
-    static std::map<int, int> g_touchIdReorderMap;
+    static std::map<intptr_t, int> g_touchIdReorderMap;
     
     static int getUnUsedIndex()
     {
@@ -235,9 +235,9 @@ const std::string& GLViewProtocol::getViewName() const
     return _viewName;
 }
 
-void GLViewProtocol::handleTouchesBegin(int num, int ids[], float xs[], float ys[])
+void GLViewProtocol::handleTouchesBegin(int num, intptr_t ids[], float xs[], float ys[])
 {
-    int id = 0;
+    intptr_t id = 0;
     float x = 0.0f;
     float y = 0.0f;
     int unusedIndex = 0;
@@ -285,9 +285,9 @@ void GLViewProtocol::handleTouchesBegin(int num, int ids[], float xs[], float ys
     dispatcher->dispatchEvent(&touchEvent);
 }
 
-void GLViewProtocol::handleTouchesMove(int num, int ids[], float xs[], float ys[])
+void GLViewProtocol::handleTouchesMove(int num, intptr_t ids[], float xs[], float ys[])
 {
-    int id = 0;
+    intptr_t id = 0;
     float x = 0.0f;
     float y = 0.0f;
     EventTouch touchEvent;
@@ -317,7 +317,7 @@ void GLViewProtocol::handleTouchesMove(int num, int ids[], float xs[], float ys[
         else
         {
             // It is error, should return.
-            CCLOG("Moving touches with id: %d error", id);
+            CCLOG("Moving touches with id: %ld error", id);
             return;
         }
     }
@@ -333,9 +333,9 @@ void GLViewProtocol::handleTouchesMove(int num, int ids[], float xs[], float ys[
     dispatcher->dispatchEvent(&touchEvent);
 }
 
-void GLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode, int num, int ids[], float xs[], float ys[])
+void GLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode, int num, intptr_t ids[], float xs[], float ys[])
 {
-    int id = 0;
+    intptr_t id = 0;
     float x = 0.0f;
     float y = 0.0f;
     EventTouch touchEvent;
@@ -370,7 +370,7 @@ void GLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode,
         } 
         else
         {
-            CCLOG("Ending touches with id: %d error", id);
+            CCLOG("Ending touches with id: %ld error", id);
             return;
         } 
 
@@ -393,12 +393,12 @@ void GLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode,
     }
 }
 
-void GLViewProtocol::handleTouchesEnd(int num, int ids[], float xs[], float ys[])
+void GLViewProtocol::handleTouchesEnd(int num, intptr_t ids[], float xs[], float ys[])
 {
     handleTouchesOfEndOrCancel(EventTouch::EventCode::ENDED, num, ids, xs, ys);
 }
 
-void GLViewProtocol::handleTouchesCancel(int num, int ids[], float xs[], float ys[])
+void GLViewProtocol::handleTouchesCancel(int num, intptr_t ids[], float xs[], float ys[])
 {
     handleTouchesOfEndOrCancel(EventTouch::EventCode::CANCELLED, num, ids, xs, ys);
 }
